Adding String to Enum conversion for hardfork names.
I needed to translate hardfork names from string to a typed enum multiple times so far when reading config, exposing JS api through N-API, etc. I think that if we keep this here, we can guarantee that this conversion will always be in sync with new hardforks that become supported by evmodin.